requirements.txt
alembic==1.13.3
ansicon==1.89.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.2.0
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
bpython==0.24
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
curtsies==0.4.2
cwcwidth==0.1.9
djlint==1.35.2
dnspython==2.6.1
EditorConfig==0.12.4
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security==5.5.2
flask-sitemapper==1.8.1
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
jsbeautifier==1.15.1
json5==0.9.25
Levenshtein==0.25.1
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pycparser==2.22
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.2
pyxdg==0.28
PyYAML==6.0.2
qrcode==7.4.2
rapidfuzz==3.9.6
razorpay==1.4.2
regex==2024.9.11
requests==2.32.3
setuptools==72.2.0
six==1.16.0
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
tabulate==0.9.0
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.31.0
uvloop==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
